Wrap particles into [0, 1] before passing to vis backend if periodic.
If the visualisation is periodic, the backend will tile the particles
as many times as needed to cover the vis region, and assumed that all
particles are in the [0, 1] normalized coordinate range to do this.
We therefore enforce that all particles are in this range (in case
the user has shifted the coordinates, for example).